// seller_backend/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        Item: {
            id: string;
            quantity: number;
        };
// ...
        CheckoutSession: components["schemas"]["CheckoutSessionBase"];
// ...
        Error: {
            /** @enum {string} */
            type: "invalid_request" | "request_not_idempotent" | "processing_error" | "service_unavailable";
            /** @description Implementation-defined error code */
            code: string;
            message: string;
            /** @description RFC 9535 JSONPath (optional) */
            param?: string;
        };
    };
    responses: never;
    parameters: {
        Authorization: string;
        ContentType: string;
        AcceptLanguage: string;
        UserAgent: string;
        IdempotencyKey: string;
        RequestId: string;
        Signature: string;
        Timestamp: string;
        APIVersion: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}